Create project
This is an easy number guessing game. In this game, players have got five chance to guess the random number between 1 to 50 which will be generated automatically. If the number guessed by the player is different than that random number then the player will get "Your guess is high" or "Your guess is too low". Then the player have to try again to meet the random number.